+++ src/components/InventoryManager/Header/AddListingStepIcons.js
