+++ src/components/InventoryManager/Header/DeleteButton.js
