+++ src/components/InventoryManager/Header/ListingSteps/ConfirmationBlock.js
