+++ src/components/InventoryManager/Header/ListingSteps/ConfirmationStep.js
