+++ src/components/InventoryManager/Header/ListingSteps/DatesStep.js
@@ -3,7 +3,7 @@ import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
 import { makeStyles } from '@material-ui/core/styles';
 import PropTypes from 'prop-types';
-import ListingInputField from '../ListingInputField';
+import ListingInputField from '../../ListingInputField';
 
 const useStyles = makeStyles({
   heading: {
