+++ src/components/InventoryManager/Header/ListingSteps/PhotoStep.js
