+++ src/components/InventoryManager/ListingsView/ListingsView.js
